Moved diffusivity and molar conductivity to transport.
